WE WERE LEAKING DA FUQIN RCON PASSWORD 💀
Not a real, serious security issue becasue docker-mc is smart(er than me) enough to rotate those for us (and the one that was „leaked” here has long since been auto-changed, even several times). Still, leaving it in here makes me look like an absolute dork, so we better remove it.

Also - apparently you're not supposed to set the RCON port here, or mc-docker's integrations will break. I guess this hasn't been a problem up until now because we were using the default port. But for certain reasons, we can't anymore (for now).